refactor: decompose ConnectTelegram
